commit 2

+++ src/main/java/com/example/TurfBookingApplication/Repository/OwnerJpaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OwnerJpaRepository extends JpaRepository<Owner, Integer> {
+public interface OwnerJpaRepository extends JpaRepository<Owner, Long> {
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
     boolean existsByPhoneNumber(String phoneNumber);
